pyproject.toml
[tool.poetry]
name = "stock-prediction"
version = "1.0.0"
description = "Enterprise Stock Price Prediction with LSTM & MLOps"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
yfinance = "^0.2.40"
pandas = "^2.2.2"
numpy = "^1.26.4"
torch = "^2.4.0"
scikit-learn = "^1.5.1"
ta-lib = {version = "^0.4.28", optional = true}
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = "^2.8.2"
pyyaml = "^6.0.2"
mlflow = "^2.16.0"
matplotlib = "^3.9.2"
plotly = "^5.22.0"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"

[tool.poetry.extras]
talib = ["ta-lib"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

